Consolidate integration test cases
They are the same tests, run through different entrypoints.
