+++ tests/Integration/SchemaDiffCommandTest.php
@@ -17,7 +17,7 @@ use Symfony\Component\Console\Tester\CommandTester;
  *       For now, this is a copy of the tests in Integration\SchemaDiffTest but run through the Command.
  *       Refactoring in future to allow mock dependencies will allow coverage of more features, e.g. ignore-databases
  */
-class SchemaDiffCommandTest extends IntegrationTestCase
+class SchemaDiffCommandTest extends SchemaDiffTestCase
 {
     private SchemaDiffCommand $command;
 
@@ -43,7 +43,7 @@ class SchemaDiffCommandTest extends IntegrationTestCase
         parent::setUp();
     }
 
-    private function runDiff(string $tableName, string &$output = null): int
+    protected function runDiff(string $tableName, string &$output = null): bool
     {
         $this->commandTester->execute([
             '--tables' => $tableName,
@@ -55,126 +55,6 @@ class SchemaDiffCommandTest extends IntegrationTestCase
 
         $output = $this->commandTester->getDisplay();
 
-        return $different;
-    }
-
-    public function testSame(): void
-    {
-        $tableName = $this->generateTableName();
-
-        $this->createTestTable(getenv('DB_DATABASE_1'), $tableName);
-        $this->createTestTable(getenv('DB_DATABASE_2'), $tableName);
-
-        $different = $this->runDiff($tableName, $output);
-
-        static::assertSame(0, $different);
-
-        static::assertEmpty($output);
-    }
-
-    public function dataSchemaOrder(): array
-    {
-        return [
-            'one-two' => [1, 2],
-            'two-one' => [2, 1],
-        ];
-    }
-
-    /**
-     * @dataProvider dataSchemaOrder
-     */
-    public function testMissingTable(int $first, int $second): void
-    {
-        $tableName = $this->generateTableName();
-
-        $this->createTestTable(getenv('DB_DATABASE_' . $first), $tableName);
-
-        $different = $this->runDiff($tableName, $output);
-
-        static::assertSame(1, $different);
-
-        $expected = "Missing table {$tableName} missing on " .
-            getenv('DB_DATABASE_' . $second) . "@{$second} exists on " . getenv('DB_DATABASE_' . $first) . "@{$first}";
-        static::assertStringStartsWith($expected, $output);
-    }
-
-    /**
-     * @dataProvider dataSchemaOrder
-     */
-    public function testMissingColumn(int $first, int $second): void
-    {
-        $tableName = $this->generateTableName();
-
-        $this->createTestTable(getenv('DB_DATABASE_' . $first), $tableName, true);
-        $this->createTestTable(getenv('DB_DATABASE_' . $second), $tableName, false);
-
-        $different = $this->runDiff($tableName, $output);
-
-        static::assertSame(1, $different);
-
-        $expected = "Missing column {$tableName}.char_col missing on " .
-            getenv('DB_DATABASE_' . $second) . "@{$second} exists on " . getenv('DB_DATABASE_' . $first) . "@{$first}";
-        static::assertStringContainsString($expected, $output);
-    }
-
-    /**
-     * @dataProvider dataSchemaOrder
-     */
-    public function testMissingIndex(int $first, int $second): void
-    {
-        $tableName = $this->generateTableName();
-
-        $this->createTestTable(getenv('DB_DATABASE_' . $first), $tableName, true, true);
-        $this->createTestTable(getenv('DB_DATABASE_' . $second), $tableName, true, false);
-
-        $different = $this->runDiff($tableName, $output);
-
-        static::assertSame(1, $different);
-
-        $expected = "Missing index {$tableName}.idx_char missing on " .
-            getenv('DB_DATABASE_' . $second) . "@{$second} exists on " . getenv('DB_DATABASE_' . $first) . "@{$first}";
-        static::assertStringStartsWith($expected, $output);
-    }
-
-    /**
-     * @dataProvider dataSchemaOrder
-     */
-    public function testDiffColumnCharset(int $first, int $second): void
-    {
-        $tableName = $this->generateTableName();
-
-        $this->createTestTable(getenv('DB_DATABASE_' . $first), $tableName, true, false, null);
-        $this->createTestTable(getenv('DB_DATABASE_' . $second), $tableName, true, false, 'utf8mb4');
-
-        $different = $this->runDiff($tableName, $output);
-
-        static::assertSame(1, $different);
-
-        $expected = "Column attribute mismatch {$tableName}.char_col attribute character set differs:";
-        static::assertStringStartsWith($expected, $output);
-
-        static::assertStringContainsString(getenv('DB_DATABASE_' . $first) . "@{$first}=ascii", $output);
-        static::assertStringContainsString(getenv('DB_DATABASE_' . $second) . "@{$second}=utf8mb4", $output);
-    }
-
-    /**
-     * @dataProvider dataSchemaOrder
-     */
-    public function testDiffTableCharset(int $first, int $second): void
-    {
-        $tableName = $this->generateTableName();
-
-        $this->createTestTable(getenv('DB_DATABASE_' . $first), $tableName, true, false, null, 'ascii');
-        $this->createTestTable(getenv('DB_DATABASE_' . $second), $tableName, true, false, null, 'utf8mb4');
-
-        $different = $this->runDiff($tableName, $output);
-
-        static::assertSame(1, $different);
-
-        $expected = "Table attribute mismatch {$tableName} attribute collation differs:";
-        static::assertStringStartsWith($expected, $output);
-
-        static::assertStringContainsString(getenv('DB_DATABASE_' . $first) . "@{$first}=ascii", $output);
-        static::assertStringContainsString(getenv('DB_DATABASE_' . $second) . "@{$second}=utf8mb4", $output);
+        return (bool)$different;
     }
 }
